+++ src/cljs-lib/src/pez_cljfmt/core.clj
@@ -1,5 +1,0 @@
-(ns pez-cljfmt.core
-  (:require [clojure.java.io :as io]))
-
-(def read-resource* (comp read-string slurp io/resource))
-(defmacro read-resource [path] `'~(read-resource* path))